+++ model/BoardCatalogResource.py
@@ -1,0 +1,7 @@
+from flask_restful import Resource
+
+from model.BoardListCatalog import BoardCatalog
+
+
+class BoardCatalogResource(BoardCatalog, Resource):
+    pass
