+++ model/BoardListResource.py
@@ -1,0 +1,7 @@
+from flask_restful import Resource
+
+from model.BoardList import BoardList
+
+
+class BoardListResource(BoardList, Resource):
+    pass
