+++ model/FirehoseResource.py
@@ -1,0 +1,7 @@
+from flask_restful import Resource
+
+from model.Firehose import Firehose
+
+
+class FirehoseResource(Firehose, Resource):
+    pass
