+++ model/Media.py
@@ -1,0 +1,6 @@
+from shared import db
+
+
+class Media(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    ext = db.Column(db.String(3), nullable=False)
