+++ model/NewPostResource.py
@@ -1,0 +1,7 @@
+from flask_restful import Resource
+
+from model.NewPost import NewPost
+
+
+class NewPostResource(NewPost, Resource):
+    pass
