+++ model/NewThreadResource.py
@@ -1,0 +1,7 @@
+from flask_restful import Resource
+
+from model.NewThread import NewThread
+
+
+class NewThreadResource(NewThread, Resource):
+    pass
