+++ model/PostReplyExtension.py
@@ -1,0 +1,18 @@
+import re
+
+from flask import url_for
+from markdown import Extension
+
+from model.PostReplyPattern import PostReplyPattern
+from model.Thread import Thread
+
+
+
+class PostReplyExtension(Extension):
+    def extendMarkdown(self, md, _):
+        post_reply = PostReplyPattern()
+        md.inlinePatterns.add("post_reply", post_reply, "_begin")
+        # modify the built-in blockquote regexp to only match
+        # when one angle bracket is present
+        # monkey patching was a mistake
+        md.parser.blockprocessors["quote"].RE = re.compile(r"(^|\n)[ ]{0,3}>(?!>)[ ]?(.*)")
