+++ model/ThreadPostsResource.py
@@ -1,0 +1,7 @@
+from flask_restful import Resource
+
+from model.ThreadPosts import ThreadPosts
+
+
+class ThreadPostsResource(ThreadPosts, Resource):
+    pass
