+++ templates/base.html
@@ -1,68 +1,84 @@
 <!doctype html>
 <html lang="en">
-	<head>
-		<!-- Required meta tags -->
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+<head>
+    <!-- Required meta tags -->
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
-		<!-- Bootstrap CSS -->
-		<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css" integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous">
+    <!-- Bootstrap CSS -->
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css"
+          integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous">
 
-		<!-- Optional JavaScript -->
-		<!-- jQuery first, then Popper.js, then Bootstrap JS -->
-		<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-		<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-		<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/js/bootstrap.min.js" integrity="sha384-smHYKdLADwkXOn1EmN1qk/HfnUcbVRZyYmZ4qpPea6sjB/pTJ0euyQp0Mk8ck+5T" crossorigin="anonymous"></script>
+    <!-- Optional JavaScript -->
+    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
+    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+            integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+            crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js"
+            integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
+            crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/js/bootstrap.min.js"
+            integrity="sha384-smHYKdLADwkXOn1EmN1qk/HfnUcbVRZyYmZ4qpPea6sjB/pTJ0euyQp0Mk8ck+5T"
+            crossorigin="anonymous"></script>
 
-		<!-- Fontawesome -->
-		<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/all.css" integrity="sha384-hWVjflwFxL6sNzntih27bfxkr27PmbbK/iSvJ+a4+0owXq79v+lsFkW54bOGbiDQ" crossorigin="anonymous">
-		{% block head %}{% endblock %}
-		<title>Maniwani - {% block title %}{% endblock %}</title>
-	</head>
-	<body>
-		<nav class="navbar navbar-expand-lg navbar-dark bg-dark sticky-top mb-2">
-			<a class="navbar-brand" href="{{ url_for("index") }}">Maniwani</a>
-			<script>
+    <!-- Fontawesome -->
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/all.css"
+          integrity="sha384-hWVjflwFxL6sNzntih27bfxkr27PmbbK/iSvJ+a4+0owXq79v+lsFkW54bOGbiDQ" crossorigin="anonymous">
+    {% block head %}{% endblock %}
+    <title>Maniwani - {% block title %}{% endblock %}</title>
+</head>
+<body>
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark sticky-top mb-2">
+    <a class="navbar-brand" href="{{ url_for(" index") }}">Maniwani</a>
+    <script>
 			 document.write('<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#site-nav"> \
 				 <span class="navbar-toggler-icon"></span> \
 				 </button>')
-			</script>
-			<div id="site-nav" class="navbar-nav">
-				{% block navbar %}{% endblock %}
-				<a href="{{ url_for("list_boards") }}" class="nav-item nav-link">Boards</a>
-				<a href="{{ url_for("slip_landing") }}" class="nav-item nav-link">Slip</a>
-				{% if board_id is not defined and board is not defined %}
-					<a href="{{ url_for("view_rules") }}" class="nav-item nav-link">Rules</a>
-				{% elif board_id is defined %}
-					<a href="{{ url_for("view_rules", board_id=board_id) }}" class="nav-item nav-link">Rules</a>
-				{% else %}
-					<a href="{{ url_for("view_rules", board_id=board.id) }}" class="nav-item nav-link">Rules</a>
-				{% endif %}
-				<a href="{{ url_for("faq") }}" class="nav-item nav-link">FAQ</a>
-				<div class="navbar-nav ml-auto">
-					{% block navbar_right %}{% endblock %}
-				</div>
-			</div>
-			<script>
+
+
+
+    </script>
+    <div id="site-nav" class="navbar-nav">
+        {% block navbar %}{% endblock %}
+        <a href="{{ url_for(" list_boards") }}" class="nav-item nav-link">Boards</a>
+        <a href="{{ url_for(" slip_landing") }}" class="nav-item nav-link">Slip</a>
+        {% if board_id is not defined and board is not defined %}
+        <a href="{{ url_for(" view_rules") }}" class="nav-item nav-link">Rules</a>
+        {% elif board_id is defined %}
+        <a href="{{ url_for(" view_rules", board_id=board_id) }}" class="nav-item nav-link">Rules</a>
+        {% else %}
+        <a href="{{ url_for(" view_rules", board_id=board.id) }}" class="nav-item nav-link">Rules</a>
+        {% endif %}
+        <a href="{{ url_for(" faq") }}" class="nav-item nav-link">FAQ</a>
+        <div class="navbar-nav ml-auto">
+            {% block navbar_right %}{% endblock %}
+        </div>
+    </div>
+    <script>
 				$("#site-nav").addClass("collapse navbar-collapse")
-			</script>
-		</nav>
-		{% with flashes = get_flashed_messages() %}
-			{% if flashes %}
-				<div class="container">
-					{% for message in flashes %}
-						<div class="alert alert-primary">
-							{{ message }}
-						</div>
-					{% endfor %}
-				</div>
-			{% endif %}
-		{% endwith %}
-		{% block content %}{% endblock %}
-		<hr>
-		<div class="container-fluid">
-			{% block footer %}{% endblock %}
-			<p class="text-center text-muted"><small>Powered by <a href="https://github.com/DangerOnTheRanger/maniwani">Maniwani</a>!</small></p>
-		</div>
-	</body>
+
+
+
+    </script>
+</nav>
+{% with flashes = get_flashed_messages() %}
+{% if flashes %}
+<div class="container">
+    {% for message in flashes %}
+    <div class="alert alert-primary">
+        {{ message }}
+    </div>
+    {% endfor %}
+</div>
+{% endif %}
+{% endwith %}
+{% block content %}{% endblock %}
+<hr>
+<div class="container-fluid">
+    {% block footer %}{% endblock %}
+    <p class="text-center text-muted">
+        <small>Powered by <a href="https://github.com/DangerOnTheRanger/maniwani">Maniwani</a>!</small>
+    </p>
+</div>
+</body>
 </html>
