+++ templates/catalog-view.html
@@ -1,5 +1,5 @@
 {% macro catalog_view(threads) %}
-	<style>
+<style>
 	 @media (max-width: 575px) {
 		 .thread, .thread-sizer {
 			 width: 98%;
@@ -30,56 +30,65 @@
 			 margin: 0.2em;
 		 }
 	 }
-	</style>
-	<script src="https://unpkg.com/imagesloaded@4/imagesloaded.pkgd.min.js"></script>
-	<script src="https://unpkg.com/masonry-layout@4/dist/masonry.pkgd.min.js"></script>
-	<div class="container-fluid">
-		<div id="grid" class="d-flex flex-wrap">
-			<script>
+
+
+
+</style>
+<script src="https://unpkg.com/imagesloaded@4/imagesloaded.pkgd.min.js"></script>
+<script src="https://unpkg.com/masonry-layout@4/dist/masonry.pkgd.min.js"></script>
+<div class="container-fluid">
+    <div id="grid" class="d-flex flex-wrap">
+        <script>
 			 document.write('<div class="thread-sizer" style="display: none;"></div>')
-			</script>
-			{% for thread in threads %}
-				{% set bg_style = "bg-light" %}
-				{% set ns = namespace(bg_style="bg-light", text_style="text-body") %}
-				{% if thread.tags %}
-					{% set ns.main_tag = thread.tags[0].name %}
-					{% if style_for_tag(ns.main_tag).get("bg_style") %}
-						{% set ns.bg_style = style_for_tag(ns.main_tag).get("bg_style") %}
-					{% endif %}
-					{% if style_for_tag(ns.main_tag).get("text_style") %}
-						{% set ns.text_style = style_for_tag(ns.main_tag).get("text_style") %}
-					{% endif %}
-				{% endif %}
-				<div class="thread border-top rounded">
-					{% if thread["media"] %}
-						<a href="{{ url_for("view_thread", thread_id=thread["id"]) }}">
-							<img class="border-left border-right w-100 rounded-top" src="{{ url_for("uploaded_thumb", media_id=thread["media"]) }}">
-						</a>
-					{% endif %}
-					{% if thread["subject"] %}
-						<div class="{{ "p-2 border-left border-right " + ns.bg_style }}">
-							<h5 class="{{ ns.text_style }}">{{ thread["subject"] }}</h5>
-						</div>
-					{% else %}
-						<div class="bg-light p-2 border-left border-right">
-							<h5 class="text-muted">No subject</h5>
-						</div>
-					{% endif %}
-					<div class="pb-2 pt-2 pl-4 pr-4 border-left border-right">
-						{{ thread["body"]|safe|truncate() }}
-					</div>
-					<div class="{{ "pt-1 border border-top-0 rounded-bottom " + ns.bg_style }}">
-						<a href="{{ url_for("view_thread", thread_id=thread["id"]) }}" class="btn btn-dark m-2">View thread</a>
-						<span class="{{ "mt-3 mr-2 float-right " + ns.text_style }}">
-							<i class="fas fa-comment"></i> {{ thread["num_replies"] }} 
-							<i class="fas fa-image"></i> {{ thread["num_media"] }}
-						</span>
-					</div>
-				</div>
-			{% endfor %}		
-		</div>
-	</div>
-	<script>
+
+
+
+        </script>
+        {% for thread in threads %}
+        {% set bg_style = "bg-light" %}
+        {% set ns = namespace(bg_style="bg-light", text_style="text-body") %}
+        {% if thread.tags %}
+        {% set ns.main_tag = thread.tags[0].name %}
+        {% if style_for_tag(ns.main_tag).get("bg_style") %}
+        {% set ns.bg_style = style_for_tag(ns.main_tag).get("bg_style") %}
+        {% endif %}
+        {% if style_for_tag(ns.main_tag).get("text_style") %}
+        {% set ns.text_style = style_for_tag(ns.main_tag).get("text_style") %}
+        {% endif %}
+        {% endif %}
+        <div class="thread border-top rounded">
+            {% if thread["media"] %}
+            <a href="{{ url_for(" view_thread", thread_id=thread["id"]) }}">
+            <img class="border-left border-right w-100 rounded-top" src="{{ url_for(" uploaded_thumb",
+            media_id=thread["media"]) }}">
+            </a>
+            {% endif %}
+            {% if thread["subject"] %}
+            <div class="{{ " p-2 border-left border-right
+            " + ns.bg_style }}">
+            <h5 class="{{ ns.text_style }}">{{ thread["subject"] }}</h5>
+        </div>
+        {% else %}
+        <div class="bg-light p-2 border-left border-right">
+            <h5 class="text-muted">No subject</h5>
+        </div>
+        {% endif %}
+        <div class="pb-2 pt-2 pl-4 pr-4 border-left border-right">
+            {{ thread["body"]|safe|truncate() }}
+        </div>
+        <div class="{{ " pt-1 border border-top-0 rounded-bottom
+        " + ns.bg_style }}">
+        <a href="{{ url_for(" view_thread", thread_id=thread["id"]) }}" class="btn btn-dark m-2">View thread</a>
+        <span class="{{ " mt-3 mr-2 float-right " + ns.text_style }}">
+        <i class="fas fa-comment"></i> {{ thread["num_replies"] }}
+        <i class="fas fa-image"></i> {{ thread["num_media"] }}
+        </span>
+    </div>
+</div>
+{% endfor %}
+</div>
+</div>
+<script>
 	 var grid = $("#grid")
 	 grid.imagesLoaded( function() {
 		 grid.masonry({
@@ -89,5 +98,8 @@
 			 horizontalOrder: true
 		 })
 	 })
-	</script>
+
+
+
+</script>
 {% endmacro %}
