+++ templates/new-thread.html
@@ -3,11 +3,11 @@
 new thread
 {% endblock %}
 {% block navbar %}
-	<a href="{{ url_for("view_catalog", board_id=board_id) }}" class="nav-item nav-link">Catalog</a>
+<a href="{{ url_for(" view_catalog", board_id=board_id) }}" class="nav-item nav-link">Catalog</a>
 {% endblock %}
 {% import "post-base.html" as post %}
 {% block content %}
-	<div class="container">
-		{{ post.post_form(url_for("thread_submit"), "board", board_id, thread_op=True) }}
-	</div>
+<div class="container">
+    {{ post.post_form(url_for("thread_submit"), "board", board_id, thread_op=True) }}
+</div>
 {% endblock %}
