+++ test/TestPost.py
@@ -1,0 +1,5 @@
+import unittest
+
+
+class TestPost(unittest.testcase):
+    pass
